+++ src/components/addPatient/index.js
@@ -1,15 +1,173 @@
-import React from 'react'
-import { View } from 'react-native'
+import React, { useState } from 'react'
+import { useDispatch } from 'react-redux'
+import { View, TouchableOpacity } from 'react-native'
+import Icon from 'react-native-vector-icons/MaterialIcons'
+import { Picker } from '@react-native-community/picker'
+import { VStack, Text, ScrollView } from 'native-base'
+import { useToast } from 'native-base'
+import * as theme from '../../constants/theme'
 import { styles } from './form.style'
-import { Body } from './body'
-import { Footer } from './footer'
-import { Header } from './header'
+import { addNewPatientList } from '../../stores/actions/user.action'
+import { setrefPatientsList } from './redux/actions'
+import { CTextinput } from '../Textinput'
+import { useDatabase } from '../../context/DatabaseContext'
+
 const AddPatient = props => {
+  const dispatch = useDispatch()
+  const ctx = useDatabase()
+  const toast = useToast()
+  const [distance] = useState(0)
+  const [sex, setSex] = useState('Sex')
+  const [name, setName] = useState('')
+  const [guardianName, setGuardianName] = useState('')
+  const [healthId, sethealthId] = useState('')
+  const [mobileNumber, setMobileNumber] = useState('')
+  const [address, setAddress] = useState('')
+  const [error_description, setErrorDescription] = useState('')
+  const { onCloseForm } = props
+  const clearForm = () => {
+    setSex('Sex')
+    setName('')
+    setGuardianName('')
+    sethealthId('')
+    setMobileNumber('')
+    setAddress('')
+    setErrorDescription('')
+  }
+
+  let body = (
+    <ScrollView showsVerticalScrollIndicator={false}>
+      <View style={styles.body}>
+        <CTextinput
+          placeholder="Name"
+          value={name}
+          onChangeText={text => setName(text)}
+        />
+        <CTextinput
+          placeholder="Guardian Name"
+          defaultValue={guardianName}
+          onChangeText={text => setGuardianName(text)}
+        />
+        <CTextinput
+          placeholder="Health id"
+          defaultValue={healthId}
+          onChangeText={text => sethealthId(text)}
+        />
+
+        <View
+          style={{
+            justifyContent: 'center',
+            flexDirection: 'row'
+          }}>
+          <View style={[styles.col, { flex: 2 }]}>
+            <CTextinput
+              placeholder="Mobile number"
+              style={{ marginLeft: 10 }}
+              defaultValue={mobileNumber}
+              onChangeText={text => setMobileNumber(text)}
+            />
+          </View>
+          <View style={[styles.col, { marginLeft: 5 }]}>
+            <View style={styles.pickerContainer}>
+              <Picker
+                value={distance}
+                selectedValue={sex}
+                onValueChange={(itemValue, itemIndex) => setSex(itemValue)}>
+                <Picker.Item label="Sex" value="Sex" />
+                <Picker.Item label="Male" value="Male" />
+                <Picker.Item label="Femal" value="Femal" />
+              </Picker>
+            </View>
+          </View>
+        </View>
+        <CTextinput
+          multiline
+          numberOfLines={4}
+          defaultValue={address}
+          onChangeText={text => setAddress(text)}
+          placeholder="Address"
+        />
+        <VStack space={1} alignItems="center" mt={3}>
+          <Text fontSize="md" color={'red.500'}>
+            {error_description}
+          </Text>
+        </VStack>
+      </View>
+    </ScrollView>
+  )
+
   return (
     <View style={styles.container}>
-      <Header />
-      <Body />
-      <Footer />
+      <View style={styles.header}>
+        <TouchableOpacity onPress={onCloseForm}>
+          <Icon
+            name="keyboard-arrow-left"
+            size={30}
+            color={theme.colors.black}
+          />
+        </TouchableOpacity>
+        <Text style={styles.headerTitle}>Add New Patient</Text>
+        <TouchableOpacity onPress={clearForm}>
+          <Text style={{ color: theme.colors.gray }}>Clear</Text>
+        </TouchableOpacity>
+      </View>
+
+      {body}
+      <View style={{ padding: 20, backgroundColor: theme.colors.white }}>
+        <TouchableOpacity
+          style={styles.btnContainer}
+          onPress={() => {
+            let data = {
+              name,
+              healthId,
+              mobileNumber,
+              sex,
+              address,
+              guardianName
+            }
+
+            if (!name) {
+              setErrorDescription('Please enter patient name !')
+            } else if (!mobileNumber) {
+              setErrorDescription('Please enter mobile number.')
+            } else if (!sex || sex === 'Sex') {
+              setErrorDescription("Selecte patient's gender")
+            } else if (!address) {
+              setErrorDescription("Enter patient's address before submitting.")
+            } else if (!guardianName) {
+              setErrorDescription("Fill patient's gaurdian's name.")
+            } else {
+              setImmediate(() => dispatch(addNewPatientList(data)))
+              ctx
+                .addPatient(
+                  name,
+                  healthId,
+                  mobileNumber,
+                  sex,
+                  address,
+                  guardianName
+                )
+                .then(res => {
+                  setImmediate(() => dispatch(setrefPatientsList(true)))
+                })
+                .then(res => {
+                  console.log('add patient err', res)
+                })
+                .catch(err => {
+                  console.log('add patient', err)
+                })
+              setImmediate(() => {
+                toast.show({
+                  title: `Patient ${name} added !! `
+                })
+                clearForm()
+                onCloseForm()
+              })
+            }
+          }}>
+          <Text style={styles.btnText}>Submit</Text>
+        </TouchableOpacity>
+      </View>
     </View>
   )
 }
