+++ src/components/addPatient/redux/actionType.js
@@ -1,13 +1,3 @@
 export const ACTION_TYPE = {
-  Submit: 'ADD_PATIENT',
-  CLEAR: 'CLEAR_FORM',
-  SET_NAME: 'SET_NAME',
-  SET_SEX: 'SET_SEX',
-  SET_GAURDIAN: 'SET_GAURDIAN',
-  SET_HEALTH_ID: 'SET_HEALTH_ID',
-  SET_MOBILE_NUMBER: 'SET_MOBILE_NUMBER',
-  SET_ADDRESS: 'SET_ADDRESS',
-  ADD_PATIENT_BUTTON_CLICKED: 'ADD_PATIENT_BUTTON_CLICKED',
-  PATIENT_FORM_EMPTY: 'PATIENT_FORM_EMPTY',
   REF_PATIENT_LIST: 'REF_PATIENT_LIST'
 }
